backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
python-jose[cryptography]
python-dotenv
pydantic
